Add code to produce the HF solutions appearing in the thesis
I've moved an old script out of /Code into /Writings/Thesis/code and
started hacking it to focus on reproducibility testing -- i.e. testing
that it produces the same results file as "hartree-results-<date>.npy"
that was used to actually produce the plots.

Also starting to do a big pylint cleanup. The initial evaluation was
-4 / 10. That's a negative 4 out of 10! It's positive now though and will
increase... :-)